+++ tests/test_decorators.py
@@ -1,14 +1,15 @@
-from django.http import HttpResponse
+from django.http import HttpRequest, HttpResponse
 
 from directory_components.decorators import skip_ga360
 
 
 @skip_ga360
-def test_function(request, *args, **kwargs):
+def get_response(request, *args, **kwargs):
     return HttpResponse()
 
 
 def test_skip_360_adds_attribute_to_response():
-    response = test_function({})
+    request = HttpRequest()
+    get_response(request)
 
-    assert response.skip_ga360 is True
+    assert request.skip_ga360 is True